renamed UNDP components to general name without UNDP
