+++ src/components/RasterLayer.svelte
@@ -12,15 +12,18 @@
   import { LayerInitialValues, TabNames } from '$lib/constants'
   import type { Layer } from '$lib/types'
   import { faChartColumn } from '@fortawesome/free-solid-svg-icons/faChartColumn'
-  import RasterHistogram from './RasterHistogram.svelte'
+  import RasterHistogram from '$components/RasterHistogram.svelte'
 
   export let layer: Layer = LayerInitialValues
+
+  $: tree = layer.tree
 
   let activeTab = ''
   let isRefinePanelVisible = false
   let isLegendPanelVisible = false
   let isOpacityPanelVisible = false
   let isHistogramPanelVisible = false
+
   $: {
     isLegendPanelVisible = false
     isRefinePanelVisible = false
@@ -44,12 +47,53 @@
     }
   }
 
-  const tabs = [
+  let tabs = [
     { label: TabNames.LEGEND, icon: faList, active: false },
     { label: TabNames.HISTOGRAM, icon: faChartColumn, active: false },
     { label: TabNames.REFINE, icon: faCalculator, active: false },
     { label: TabNames.OPACITY, icon: faDroplet, active: false },
   ]
+
+  $: {
+    if (tree && tree.isMosaicJSON) {
+      // disable other menus since they are not working for mosaicjson layer currently
+      tabs = [{ label: TabNames.OPACITY, icon: faDroplet, active: false }]
+    }
+  }
+
+  const handleKeyDown = (event: KeyboardEvent) => {
+    if (event.key === 'ArrowLeft') {
+      setLeftActiveTab(activeTab)
+    }
+    if (event.key === 'ArrowRight') {
+      setRightActiveTab(activeTab)
+    }
+  }
+
+  const setLeftActiveTab = (currentActiveTab: string) => {
+    const currentTabIndex = tabs.findIndex((tab) => tab.label === currentActiveTab)
+    const nextTabIndex = currentTabIndex - 1
+    if (nextTabIndex < 0) {
+      activeTab = tabs[tabs.length - 1].label
+      document.getElementById(`${activeTab}-${layer.definition.id}`)?.focus()
+    } else {
+      activeTab = tabs[nextTabIndex].label
+      document.getElementById(`${activeTab}-${layer.definition.id}`)?.focus()
+    }
+  }
+
+  const setRightActiveTab = (currentActiveTab: string) => {
+    const currentTabIndex = tabs.findIndex((tab) => tab.label === currentActiveTab)
+    const nextTabIndex = currentTabIndex + 1
+    const nextTab = tabs[nextTabIndex]
+    if (nextTab) {
+      activeTab = nextTab.label
+      document.getElementById(`${activeTab}-${layer.definition.id}`)?.focus()
+    } else {
+      activeTab = tabs[0].label
+      document.getElementById(`${activeTab}-${layer.definition.id}`)?.focus()
+    }
+  }
 </script>
 
 <div class="raster-layer-container" transition:fade>
@@ -57,19 +101,25 @@
     <p class="panel-heading">
       <LayerNameGroup {layer} />
     </p>
-    <p class="panel-tabs">
-      {#each tabs as tab}
-        <a
-          href={'#'}
-          on:click={() => (activeTab === tab.label ? (activeTab = '') : (activeTab = tab.label))}
-          class={activeTab === tab.label ? 'is-active' : ''}>
-          <span>
-            <Fa icon={tab.icon} size="sm" />
-          </span>
-          {tab.label}
-        </a>
+    <ul class="panel-tabs" role="tablist" tabindex="0">
+      {#each tabs as tab, index}
+        <li>
+          <a
+            role="tab"
+            aria-label={tab.label}
+            id={`${tab.label}-${layer.definition.id}`}
+            on:keydown={handleKeyDown}
+            href={'#'}
+            on:click={() => (activeTab === tab.label ? (activeTab = '') : (activeTab = tab.label))}
+            class={activeTab === tab.label ? 'is-active' : ''}>
+            <span>
+              <Fa icon={tab.icon} size="sm" />
+            </span>
+            {tab.label}
+          </a>
+        </li>
       {/each}
-    </p>
+    </ul>
 
     <p class="panel-content">
       {#if isLegendPanelVisible === true}
@@ -88,22 +138,32 @@
 </div>
 
 <style lang="scss">
+  $gray-700: #232e3d;
+  $dark-red: #d12800;
+
+  .is-active {
+    border-bottom: 2px solid $dark-red !important;
+  }
   .raster-layer-container {
     margin-left: 15px;
     margin-bottom: 20px;
 
     .panel-tabs {
       padding-top: 10px;
+      border: none;
 
       a {
         margin-right: 5px;
+        font-weight: bold;
+        text-transform: capitalize;
+        color: $gray-700;
+        font-family: ProximaNova, sans-serif;
 
         span {
           margin-right: 3px;
         }
       }
     }
-
     .panel-content {
       padding: 10px;
       padding-top: 15px;
